backend/requirements.txt
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djoser==2.1.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
django-colorfield==0.8.0
django-cors-headers==3.11.0
reportlab==3.6.12
django-extra-fields==3.0.2